rename to AppInsightsClient
